+++ src/Sylius/Bundle/CartBundle/spec/Sylius/Bundle/CartBundle/Model/CartItemSpec.php
@@ -9,20 +9,18 @@
  * file that was distributed with this source code.
  */
 
-namespace spec\Sylius\Bundle\CartBundle\Entity;
+namespace spec\Sylius\Bundle\CartBundle\Model;
 
-use PHPSpec2\ObjectBehavior;
+use PhpSpec\ObjectBehavior;
 
 /**
- * Cart item entity spec.
- *
  * @author Paweł Jędrzejewski <pjedrzejewski@diweb.pl>
  */
-class CartItem extends ObjectBehavior
+class CartItemSpec extends ObjectBehavior
 {
     function it_is_initializable()
     {
-        $this->shouldHaveType('Sylius\Bundle\CartBundle\Entity\CartItem');
+        $this->shouldHaveType('Sylius\Bundle\CartBundle\Model\CartItem');
     }
 
     function it_implements_Sylius_cart_item_interface()
@@ -30,8 +28,8 @@ class CartItem extends ObjectBehavior
         $this->shouldImplement('Sylius\Bundle\CartBundle\Model\CartItemInterface');
     }
 
-    function it_extends_Sylius_cart_item_model()
+    function it_extends_Sylius_order_item()
     {
-        $this->shouldHaveType('Sylius\Bundle\CartBundle\Model\CartItem');
+        $this->shouldHaveType('Sylius\Bundle\SalesBundle\Model\OrderItem');
     }
 }