+++ src/Sylius/Bundle/CartBundle/spec/Sylius/Bundle/CartBundle/Provider/CartProviderSpec.php
@@ -11,19 +11,17 @@
 
 namespace spec\Sylius\Bundle\CartBundle\Provider;
 
-use PHPSpec2\ObjectBehavior;
+use PhpSpec\ObjectBehavior;
 
 /**
- * Cart provider spec.
- *
  * @author Paweł Jędrzejewski <pjedrzejewski@diweb.pl>
  */
-class CartProvider extends ObjectBehavior
+class CartProviderSpec extends ObjectBehavior
 {
     /**
-     * @param Sylius\Bundle\CartBundle\Storage\CartStorageInterface $storage
-     * @param Doctrine\Common\Persistence\ObjectManager             $manager
-     * @param Doctrine\Common\Persistence\ObjectRepository          $repository
+     * @param Sylius\Bundle\CartBundle\Storage\CartStorageInterface  $storage
+     * @param Doctrine\Common\Persistence\ObjectManager              $manager
+     * @param Sylius\Bundle\ResourceBundle\Model\RepositoryInterface $repository
      */
     function let($storage, $manager, $repository)
     {